Bump rocksdb to 0.21
Fixes #315

Build of sonic-server is failing on NixOS
for a few weeks now, see https://hydra.nixos.org/build/239524816

This failing build has been caused by a clang 16 upgrade,
because rocksdb depends on a old rust-bindgen version which
is not compatible with this version of clang.

See https://github.com/rust-lang/rust-bindgen/issues/2312

A resolution is to upgrade rocksdb to its latest version.